lightclient sync test mod
