+++ zingolib/src/lightclient/sync.rs
@@ -665,3 +665,42 @@ impl LightClient {
         response
     }
 }
+
+#[cfg(all(test, feature = "testvectors"))]
+pub mod test {
+    use crate::{
+        lightclient::LightClient,
+        testutils::chain_generics::{
+            conduct_chain::ConductChain as _, live_chain::LiveChain, with_assertions,
+        },
+        wallet::disk::testing::examples::{
+            ExampleMSKMGDBHOTBPETCJWCSPGOPPVersion, ExampleTestnetWalletSeed, ExampleWalletNetwork,
+        },
+    };
+
+    pub(crate) async fn sync_example_wallet(wallet_case: ExampleWalletNetwork) -> LightClient {
+        std::env::set_var("RUST_BACKTRACE", "1");
+        let wallet = wallet_case.load_example_wallet().await;
+        let lc = LightClient::create_from_wallet_async(wallet).await.unwrap();
+        let _ = lc.do_sync(true).await;
+        lc
+    }
+
+    /// this is a live sync test. its execution time scales linearly since last updated
+    #[tokio::test]
+    async fn testnet_sync_mskmgdbhotbpetcjwcspgopp_latest() {
+        sync_example_wallet(ExampleWalletNetwork::Testnet(
+            ExampleTestnetWalletSeed::MSKMGDBHOTBPETCJWCSPGOPP(
+                ExampleMSKMGDBHOTBPETCJWCSPGOPPVersion::Ga74fed621,
+            ),
+        ))
+        .await;
+    }
+    /// this is a live sync test. its execution time scales linearly since last updated
+    #[tokio::test]
+    async fn testnet_sync_cbbhrwiilgbrababsshsmtpr_latest() {
+        sync_example_wallet(ExampleWalletNetwork::Testnet(
+                ExampleTestnetWalletSeed::CBBHRWIILGBRABABSSHSMTPR(crate::wallet::disk::testing::examples::ExampleCBBHRWIILGBRABABSSHSMTPRVersion::G2f3830058)            ))
+            .await;
+    }
+}
